requirements.txt
python-slugify
tqdm
